reorganize delegate files
