+++ packages/delegate/src/results/memoize.ts
