+++ packages/delegate/src/results/mergeFields.ts
@@ -1,8 +1,9 @@
 import { FieldNode, SelectionNode, Kind, GraphQLResolveInfo, SelectionSetNode } from 'graphql';
 
-import { mergeProxiedResults } from './proxiedResult';
-import { MergedTypeInfo, SubschemaConfig } from './types';
+import { MergedTypeInfo, SubschemaConfig } from '../types';
+
 import { memoize3, memoize2 } from './memoize';
+import { mergeProxiedResults } from './mergeProxiedResults';
 
 const sortSubschemasByProxiability = memoize3(function (
   mergedTypeInfo: MergedTypeInfo,
